+++ packages/nextjs/hooks/gyro/index.ts
@@ -2,4 +2,4 @@ export * from "./useEclpPoolChart";
 export * from "./useGetECLPLiquidityProfile";
 export * from "./drawLiquidityECLP";
 export * from "./useValidateEclpParams";
-export * from "./useEclpPoolSpotPrice";
+export * from "./useAutofillStarterParams";
